+++ src/libpmemblk/blk.c
@@ -600,9 +600,9 @@ pmemblk_nblock(PMEMblkpool *pbp)
  * pmemblk_read -- read a block in a block memory pool
  */
 int
-pmemblk_read(PMEMblkpool *pbp, void *buf, off_t blockno)
+pmemblk_read(PMEMblkpool *pbp, void *buf, long long blockno)
 {
-	LOG(3, "pbp %p buf %p blockno %lld", pbp, buf, (long long)blockno);
+	LOG(3, "pbp %p buf %p blockno %lld", pbp, buf, blockno);
 
 	if (blockno < 0) {
 		ERR("negative block number");
@@ -625,9 +625,9 @@ pmemblk_read(PMEMblkpool *pbp, void *buf, off_t blockno)
  * pmemblk_write -- write a block (atomically) in a block memory pool
  */
 int
-pmemblk_write(PMEMblkpool *pbp, const void *buf, off_t blockno)
+pmemblk_write(PMEMblkpool *pbp, const void *buf, long long blockno)
 {
-	LOG(3, "pbp %p buf %p blockno %lld", pbp, buf, (long long)blockno);
+	LOG(3, "pbp %p buf %p blockno %lld", pbp, buf, blockno);
 
 	if (pbp->rdonly) {
 		ERR("EROFS (pool is read-only)");
@@ -656,9 +656,9 @@ pmemblk_write(PMEMblkpool *pbp, const void *buf, off_t blockno)
  * pmemblk_set_zero -- zero a block in a block memory pool
  */
 int
-pmemblk_set_zero(PMEMblkpool *pbp, off_t blockno)
+pmemblk_set_zero(PMEMblkpool *pbp, long long blockno)
 {
-	LOG(3, "pbp %p blockno %lld", pbp, (long long)blockno);
+	LOG(3, "pbp %p blockno %lld", pbp, blockno);
 
 	if (pbp->rdonly) {
 		ERR("EROFS (pool is read-only)");
@@ -687,9 +687,9 @@ pmemblk_set_zero(PMEMblkpool *pbp, off_t blockno)
  * pmemblk_set_error -- set the error state on a block in a block memory pool
  */
 int
-pmemblk_set_error(PMEMblkpool *pbp, off_t blockno)
+pmemblk_set_error(PMEMblkpool *pbp, long long blockno)
 {
-	LOG(3, "pbp %p blockno %lld", pbp, (long long)blockno);
+	LOG(3, "pbp %p blockno %lld", pbp, blockno);
 
 	if (pbp->rdonly) {
 		ERR("EROFS (pool is read-only)");
